requirements.bytewax.txt
bytewax==0.16.2
confluent-kafka
